requirements.txt
#
# Copyright (C) 2021 Vaticle
#
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
#

absl-py==0.11.0
astor==0.8.1
behave==1.2.6
cached-property==1.5.2
cloudpickle==1.1.1
contextlib2==0.6.0.post1
cycler==0.10.0
decorator==4.4.2
dm-sonnet==1.35
dm-tree==0.1.5
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
typedb-client==2.1.0
graph-nets==1.0.4
grpcio==1.35.0
h5py==3.1.0
importlib-metadata==3.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.1.1
networkx==2.5
numpy==1.19.5
pandas==0.25.1
parse==1.18.0
parse-type==0.5.2
protobuf==3.14.0
PyHamcrest==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
scipy==1.3.1
semantic-version==2.8.5
six==1.15.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-probability==0.7.0
termcolor==1.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
